+++ packages/remax-cli/src/builtinPlugins.ts
@@ -1,6 +1,6 @@
 import errorScreen from '@remax/plugin-error-screen';
 import devtools from '@remax/plugin-devtools';
-import { Options, Platform, Plugin } from '@remax/types';
+import type { Options, Plugin } from '@remax/types';
 
 export const builtinPlugins = (
   options: Options
@@ -15,7 +15,7 @@ export const builtinPlugins = (
       init: errorScreen,
     },
   ];
-  if (options.target !== Platform.web && isDev && options.devtools) {
+  if (options.target !== 'web' && isDev && options.devtools) {
     plugins.push({
       optionKey: 'devtools',
       init: devtools,