+++ packages/remax-cli/src/extensions.ts
@@ -1,4 +1,4 @@
-import { Platform } from '@remax/types';
+import type { Platform } from '@remax/types';
 
 const extensions = ['.mjs', '.js', '.jsx', '.json', '.ts', '.tsx'];
 
@@ -10,7 +10,7 @@ export const rename = (file: string, ext = '.js') => file.replace(matcher, ext);
 
 export const targetExtensions = (target: Platform) => {
   let ext = extensions.map(ext => `.${target}${ext}`);
-  if (target !== Platform.web) {
+  if (target !== 'web') {
     ext = ext.concat(extensions.map(ext => `.mini${ext}`));
   }
   return ext.concat(extensions);