+++ packages/remax-cli/src/index.ts
@@ -1,9 +1,8 @@
 import yargs from 'yargs';
-import { Options } from '@remax/types';
+import type { Options } from '@remax/types';
 import { internalBuildApp, buildMiniPlugin } from './build';
 import getConfig from './getConfig';
 import API from './API';
-import output from './build/utils/output';
 
 export * from './legacyExport';
 
@@ -75,10 +74,14 @@ export default class RemaxCLI {
               alias: 'd',
               type: 'boolean',
               default: true,
+            })
+            .option('loglevel', {
+              describe: '展示日志级别',
+              type: 'string',
+              default: 'verbose',
             });
         },
         (argv: any) => {
-          output.message('🚀 开始构建\n', 'blue');
           internalBuildApp({ ...this.options, ...argv }, this.api!);
           try {
             require('remax-stats').run({ type: 'remax' });