+++ packages/remax-runtime/src/hostConfig/diffProperties.ts
@@ -1,4 +1,4 @@
-const STYLE = 'style';
+const STYLE = ['style', 'placeholderStyle'];
 const CHILDREN = 'children';
 const CLASS_NAME = 'className';
 
@@ -6,51 +6,50 @@ export default function diffProperties(
   lastRawProps: Record<string, any> | null | undefined,
   nextRawProps: Record<string, any> | null | undefined
 ): null | any[] {
-  let updatePayload: null | any[] = null;
+  const updatePayload: any[] = [];
 
   const lastProps: any = lastRawProps;
   const nextProps: any = nextRawProps;
 
   let propKey: string;
   let styleName: string;
-  let styleUpdates: Record<string, any> | null = null;
+
+  const hasOwnProperty = Object.prototype.hasOwnProperty;
+
+  const styleUpdates: Record<string, Record<string, any>> = {};
 
   for (propKey in lastProps) {
     if (
-      Object.prototype.hasOwnProperty.call(nextProps, propKey) ||
-      !Object.prototype.hasOwnProperty.call(lastProps, propKey) ||
+      hasOwnProperty.call(nextProps, propKey) ||
+      !hasOwnProperty.call(lastProps, propKey) ||
       lastProps[propKey] == null
     ) {
       continue;
     }
-    if (propKey === STYLE) {
+    if (STYLE.includes(propKey)) {
       const lastStyle = lastProps[propKey];
       for (styleName in lastStyle) {
-        if (Object.prototype.hasOwnProperty.call(lastStyle, styleName)) {
-          if (!styleUpdates) {
-            styleUpdates = {};
+        if (hasOwnProperty.call(lastStyle, styleName)) {
+          if (!styleUpdates[propKey]) {
+            styleUpdates[propKey] = {};
           }
-          styleUpdates[styleName] = '';
+          styleUpdates[propKey][styleName] = '';
         }
       }
     } else {
       // For all other deleted properties we add it to the queue. We use
       // the whitelist in the commit phase instead.
-      (updatePayload = updatePayload || []).push(propKey, propKey === CLASS_NAME ? '' : null);
+      updatePayload.push(propKey, propKey === CLASS_NAME ? '' : null);
     }
   }
 
   for (propKey in nextProps) {
     const nextProp = nextProps[propKey];
     const lastProp = lastProps != null ? lastProps[propKey] : undefined;
-    if (
-      !Object.prototype.hasOwnProperty.call(nextProps, propKey) ||
-      nextProp === lastProp ||
-      (nextProp == null && lastProp == null)
-    ) {
+    if (!hasOwnProperty.call(nextProps, propKey) || nextProp === lastProp || (nextProp == null && lastProp == null)) {
       continue;
     }
-    if (propKey === STYLE) {
+    if (STYLE.includes(propKey)) {
       if (process.env.NODE_ENV === 'development') {
         if (nextProp) {
           // Freeze the next style object so that we can assume it won't be
@@ -61,55 +60,47 @@ export default function diffProperties(
       if (lastProp) {
         // Unset styles on `lastProp` but not on `nextProp`.
         for (styleName in lastProp) {
-          if (
-            Object.prototype.hasOwnProperty.call(lastProp, styleName) &&
-            (!nextProp || !Object.prototype.hasOwnProperty.call(nextProp, styleName))
-          ) {
-            if (!styleUpdates) {
-              styleUpdates = {};
+          if (hasOwnProperty.call(lastProp, styleName) && (!nextProp || !hasOwnProperty.call(nextProp, styleName))) {
+            if (!styleUpdates[propKey]) {
+              styleUpdates[propKey] = {};
             }
-            styleUpdates[styleName] = '';
+            styleUpdates[propKey][styleName] = '';
           }
         }
         // Update styles that changed since `lastProp`.
         for (styleName in nextProp) {
-          if (
-            Object.prototype.hasOwnProperty.call(nextProp, styleName) &&
-            lastProp[styleName] !== nextProp[styleName]
-          ) {
-            if (!styleUpdates) {
-              styleUpdates = {};
+          if (hasOwnProperty.call(nextProp, styleName) && lastProp[styleName] !== nextProp[styleName]) {
+            if (!styleUpdates[propKey]) {
+              styleUpdates[propKey] = {};
             }
-            styleUpdates[styleName] = nextProp[styleName];
+            styleUpdates[propKey][styleName] = nextProp[styleName];
           }
         }
       } else {
         // Relies on `updateStylesByID` not mutating `styleUpdates`.
-        if (!styleUpdates) {
-          if (!updatePayload) {
-            updatePayload = [];
-          }
-          updatePayload.push(propKey, styleUpdates);
+        if (!styleUpdates[propKey]) {
+          updatePayload.push(propKey, null);
         }
-        styleUpdates = nextProp;
+        styleUpdates[propKey] = nextProp;
       }
     } else if (propKey === CHILDREN) {
       if (lastProp !== nextProp && (typeof nextProp === 'string' || typeof nextProp === 'number')) {
-        (updatePayload = updatePayload || []).push(propKey, '' + nextProp);
+        updatePayload.push(propKey, '' + nextProp);
       }
     } else {
       // For any other property we always add it to the queue and then we
       // filter it out using the whitelist during the commit.
-      (updatePayload = updatePayload || []).push(propKey, nextProp);
+      updatePayload.push(propKey, nextProp);
     }
   }
-  if (styleUpdates) {
-    // 由于 style 要转换成 string， 所以必须整个 style 对象都更新
-    (updatePayload = updatePayload || []).push(STYLE, {
-      ...lastProps[STYLE],
-      ...styleUpdates,
-    });
+
+  // 由于 style 要转换成 string， 所以必须整个 style 对象都更新
+  for (const styleKey in styleUpdates) {
+    const styleValue = styleUpdates[styleKey];
+    if (styleValue) {
+      updatePayload.push(styleKey, Object.assign({}, lastProps[styleKey], styleValue));
+    }
   }
 
-  return updatePayload;
+  return updatePayload.length ? updatePayload : null;
 }