+++ CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.3](https://github.com/remaxjs/remax/compare/v2.11.2...v2.11.3) (2020-12-31)
+
+### Bug Fixes
+
+- **wechat:** 修复 ScrollView 组件缺失的属性 ([#1460](https://github.com/remaxjs/remax/issues/1460)) ([be3c45d](https://github.com/remaxjs/remax/commit/be3c45dbcf7980c7ff7bc69f3922500abb085e67))
+
+## [2.11.2](https://github.com/remaxjs/remax/compare/v2.11.1...v2.11.2) (2020-12-17)
+
+### Bug Fixes
+
+- 修复 createHostComponent 无法正确注册组件的问题 ([7ca4938](https://github.com/remaxjs/remax/commit/7ca4938998450ba1a1d5d61955ff37628fd2cd28))
+
+## [2.11.1](https://github.com/remaxjs/remax/compare/v2.11.0...v2.11.1) (2020-12-17)
+
+### Bug Fixes
+
+- 修复 style props 变更的错误异常 ([#1446](https://github.com/remaxjs/remax/issues/1446)) ([44d43a7](https://github.com/remaxjs/remax/commit/44d43a75cba47fcae7b64085c52ea858fe017a83))
+
+# [2.11.0](https://github.com/remaxjs/remax/compare/v2.10.1...v2.11.0) (2020-12-14)
+
+### Bug Fixes
+
+- pageEvent 新增 onShareTimeline 类型 ([#1425](https://github.com/remaxjs/remax/issues/1425)) ([02558e2](https://github.com/remaxjs/remax/commit/02558e2c1a2b4f8c8b555f4063ce471ea0623222))
+- 修复 placeholderStyle 属性 diff 错误的问题 ([#1429](https://github.com/remaxjs/remax/issues/1429)) ([b7622a4](https://github.com/remaxjs/remax/commit/b7622a4ff381f0a3eab113fab99f983dc0247122))
+
+### Features
+
+- build cli 增加 --loglevel 参数 ([#1435](https://github.com/remaxjs/remax/issues/1435)) ([a207091](https://github.com/remaxjs/remax/commit/a20709168f4a30dec5e4b759421503baddc6b18f))
+
 ## [2.10.1](https://github.com/remaxjs/remax/compare/v2.10.0...v2.10.1) (2020-12-06)
 
 ### Bug Fixes