+++ packages/remax-macro/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.11.3](https://github.com/remaxjs/remax/compare/v2.11.2...v2.11.3) (2020-12-31)
+
+**Note:** Version bump only for package @remax/macro
+
+## [2.11.2](https://github.com/remaxjs/remax/compare/v2.11.1...v2.11.2) (2020-12-17)
+
+### Bug Fixes
+
+- 修复 createHostComponent 无法正确注册组件的问题 ([7ca4938](https://github.com/remaxjs/remax/commit/7ca4938998450ba1a1d5d61955ff37628fd2cd28))
+
+## [2.11.1](https://github.com/remaxjs/remax/compare/v2.11.0...v2.11.1) (2020-12-17)
+
+**Note:** Version bump only for package @remax/macro
+
+# [2.11.0](https://github.com/remaxjs/remax/compare/v2.10.1...v2.11.0) (2020-12-14)
+
+### Bug Fixes
+
+- pageEvent 新增 onShareTimeline 类型 ([#1425](https://github.com/remaxjs/remax/issues/1425)) ([02558e2](https://github.com/remaxjs/remax/commit/02558e2c1a2b4f8c8b555f4063ce471ea0623222))
+
 ## [2.10.1](https://github.com/remaxjs/remax/compare/v2.10.0...v2.10.1) (2020-12-06)
 
 ### Bug Fixes